docs+style(streaming): scope layer-4 doc.go to daemon/ops; fix lint
doc.go: add the Daemon (startup/daemon/config_validate) and Operability
(recovery/audit/audit_invariants) groups; this completes Slice 1, so the
note now points only at Slices 2 & 3 (events, tx-hash).

golangci-lint (this layer's own new findings) — auto-fixable ones via
golangci-lint --fix (protogetter, intrange, modernize, gofumpt, misspell,
perfsprint, some testifylint) plus:
- exhaustive: handle StateFrozen in the INV-2 switch (the quiescent state)
- testifylint: assert.Equal(float64,...) -> assert.InDelta(...,0) (14)
- revive: rename unused params (buildProductionBoundaries ctx,
  newCaptiveCoreOpener logger) to _
- unconvert: drop redundant uint32(c0First)
- lll: move long t.Parallel() trailing comments above the call; wrap a
  ProcessConfig literal
- nonamedreturns: drop names on snapshotLag; //nolint on the test helpers
  whose named outputs document the returned pair/handles
- //nolint for intentional/forward/complex spots: nilerr (ctx-canceled is
  a clean shutdown), nestif x2, contextcheck (test), unparam
  (newCaptiveCoreOpener stub until #772; test-helper params that vary in
  later slices), gocognit/cyclop (auditDiskMatchesMeta)

+++ cmd/stellar-rpc/internal/fullhistory/streaming/e2e_test.go
@@ -52,7 +52,7 @@ import (
 // e2eGetter is the FAKE captive-core ledger getter: a resumable LedgerGetter the
 // ingestion loop polls by sequence (the design's core.GetLedger(ctx, seq)). It
 // returns the frame for the requested seq when it has one, and once the poll
-// runs past the synthetic backlog it blocks until ctx is cancelled (a live tip
+// runs past the synthetic backlog it blocks until ctx is canceled (a live tip
 // stream ends only on shutdown). It records the FIRST seq it was asked for so
 // the restart step can assert the daemon re-derived the watermark and resumed
 // with no gap.
@@ -145,6 +145,8 @@ format = "text"
 // CANNOT open a second handle on the same path while the daemon runs — instead
 // it reads durable state through the daemon's own catalog, which is safe for
 // concurrent reads.
+//
+//nolint:nonamedreturns // named outputs label the (cancel, done, catalog) handles
 func runDaemonInBackground(
 	t *testing.T, cfgPath string, core *e2eCore, served *atomic.Int32, metrics Metrics,
 ) (cancel context.CancelFunc, done <-chan error, catCh <-chan *Catalog) {
@@ -217,6 +219,8 @@ func waitClean(t *testing.T, cancel context.CancelFunc, done <-chan error) {
 //	finish with Catalog.Audit → Clean.
 //
 // Correctness is asserted at every step.
+//
+//nolint:funlen // full lifecycle E2E with assertions at every step
 func TestE2E_DaemonLifecycle_FirstStartIngestFreezeRestartPrune(t *testing.T) {
 	if testing.Short() {
 		t.Skip("e2e ingests a full 10k-ledger chunk; skipped in -short")
@@ -288,7 +292,7 @@ func TestE2E_DaemonLifecycle_FirstStartIngestFreezeRestartPrune(t *testing.T) {
 	}, 60*time.Second, 50*time.Millisecond, "the boundary ticks must freeze+discard chunks 0 and 1")
 
 	require.GreaterOrEqual(t, served.Load(), int32(1), "reads were served")
-	require.Equal(t, uint32(c0First), core.resumeSeen.Load(),
+	require.Equal(t, c0First, core.resumeSeen.Load(),
 		"first start resumes captive core at genesis (watermark+1)")
 
 	// --- Correctness: chunks 0 and 1 ledger cold artifacts froze and exist on disk. ---
